Split Region into Region and DataDrivenRegion
